Reorganize package code (#201)
