+++ package/src/index.ts
@@ -1,24 +1,10 @@
-export { activateCodeIntel, LSPProviders } from './activation'
+export { activateCodeIntel } from './activation'
+export { HandlerArgs } from './search/handler'
+export { CommentStyle, BlockCommentStyle } from './search/comments'
+export { LSPProviders } from './lsp/providers'
 export {
     AbortError,
     createAbortError,
     isAbortError,
     throwIfAbortError,
 } from './abort'
-export { impreciseBadge } from './badges'
-export { Handler, HandlerArgs } from './handler'
-export {
-    initLSIF,
-    asyncFirst,
-    asyncWhen,
-    when,
-    wrapMaybe,
-    Maybe,
-    MaybeProviders,
-    noopMaybeProviders,
-    mkIsLSIFAvailable,
-    hover,
-    definition,
-    references,
-    Providers,
-} from './lsif'
