+++ package/src/search/handler.test.ts
@@ -7,7 +7,7 @@ import {
     findSearchToken,
 } from './handler'
 import { TextDocument } from 'sourcegraph'
-import { pythonStyle, cStyle } from '../../languages'
+import { pythonStyle, cStyle } from '../../../languages'
 import { createStubTextDocument } from '@sourcegraph/extension-api-stubs'
 
 describe('search requests', () => {
