+++ package/src/search/providers.ts
@@ -1,0 +1,18 @@
+import * as sourcegraph from 'sourcegraph'
+
+export interface SearchProviders {
+    definition: (
+        doc: sourcegraph.TextDocument,
+        pos: sourcegraph.Position
+    ) => Promise<sourcegraph.Definition>
+
+    references: (
+        doc: sourcegraph.TextDocument,
+        pos: sourcegraph.Position
+    ) => Promise<sourcegraph.Location[] | null>
+
+    hover: (
+        doc: sourcegraph.TextDocument,
+        pos: sourcegraph.Position
+    ) => Promise<sourcegraph.Hover | null>
+}
