+++ package/src/search/settings.ts
@@ -1,0 +1,7 @@
+/**
+ * @see package.json contributes.configuration section for the configuration schema.
+ */
+export interface Settings {
+    ['basicCodeIntel.debug.traceSearch']?: boolean
+    ['fileLocal']?: boolean
+}
